pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fieldworkarena"
version = "0.1.0"
description = "Field Work Arena"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.5",
    "google-adk>=1.14.1",
    "google-genai>=1.36.0",
    "nltk>=3.9.2",
    "openai>=2.7.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "litellm>=1.74",
    "pypdf>=6.4.0",
    "pillow>=12.0",
    "opencv-python>=4.12",
    "huggingface-hub>=1.2.3"
]

[project.scripts]
fwa-run = "fieldworkarena.run_scenario:main"
fwa-server = "fieldworkarena.agent.fwa_green_agent:main"

[dependency-groups]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.18.1",
    "pre-commit>=3.4.0",
    "bandit>=1.7.5",
    "safety>=2.3.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/fieldworkarena"]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | deployment
    | data
    | logs
)/
'''

[tool.ruff]
target-version = "py312"
line-length = 100  # Blackと統一

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG001", # unused-function-argument
    "SIM118", # key-in-dict
]
ignore = [
    "E402",   # module-level-import-not-at-top-of-file
    "E722",   # bare-except
    "F811",   # redefined-while-unused
    "B008",   # do-not-call-argument-defaults
    "B904",   # raise-without-from-inside-except
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "deployment",
    "data",
    "logs",
    "docs",
    "*.egg-info",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101", "ARG001", "ARG002"]

[tool.ruff.lint.isort]
known-first-party = ["fieldworkarena"]
force-sort-within-sections = true
split-on-trailing-comma = true
combine-as-imports = true
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder"
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
exclude = [
    "deployment/",
    "data/",
    "logs/",
    "build/",
    "dist/",
]

[[tool.mypy.overrides]]
module = [
    "autogen.*",
    "browsergym.*",
    "a2a.*",
    "langgraph.*",
    "langchain.*",
    "osmnx.*",
    "google_adk.*",
    "mcp.*",
    "fastmcp.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "asyncio: marks tests as async tests",
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
asyncio_mode = "auto"